Extend support for if statements; allow in list checks
Updated grammars for levels 4 - 7 to accept both `foo is bar` and `foo in bar` as conditions for the if-statement. In the latter case `bar` should be a list, because generating lists in place is not supported.

Also the implementation of `ifelse` has been moved so it is closer to the implementation of `if`.