fix: fix uncaught exception on level 4
If you run the following code on level 4:

```
people is mom, dad, Emma, Sophie
dishwasher is people at random
if dishwasher is Sophie print ... too bad I have to do the dishes ... else print 'luckily no dishes because' ... 'is already washing up'
```

The following error appears:

> We can't run your program.
>
> 'list' object has no attribute 'find'

This is because the second return value of `IsValid()`,
which is expected to be a string, is all of a sudden
a list of strings. In particular, it's the following list of
strings:

```
['print without quotes', 'print without quotes']
```

This value is passed to `closest_command`, which explodes because
it was expecting a string and not a list, leading to the **list has no
find** error.

Normalize unexpected list back to a string by just taking the first
element.

Also found some accidental nonprintable characters in the YAML, which
I removed.