chore: add `pylint` validation (#440)
Add `pylint` into the standard set of tests. `pylint` is a style
checker for Python. There are currently many style issues with the code
base, so the suite of generic style checking has been disabled. Instead,
we just check for obvious errors (misspellings of variables, etc) and
throw an error if we find any.

`pylint` found 3 in the current code base:

- Two functions in a testing class that had the same name (so the
  2nd function definition would overwrite the first). Removed the first
  test, since it was trying forms that were illegal at level 2 anyway, so
  could never work.
- Illegal access of `.line`, `.column`, `.allowed` on an instance of
  something that's declared as `Exception`. Some investigation teaches
  that the only Lark exception that has all 3 of these attributes is
  `UnexpectedCharacters`, so changed to catching only that exception.
- Some issues where we were assigning to fields of `g` (`g.prefix = ...`),
  which is allowed by pylint couldn't determine that from available declarations.
  Upgrading Flask from `1.1.1` -> `1.1.4` fixed that.

Added scripts to do all of the validations we check:

```
build-tools/validate-python
build-tools/validate-tests
build-tools/validate-yaml
```

And a single script:

```
build-tools/validate
```

To run all validations.

All validations are now run (individually) in the PR validation workflow
script as well.

Changing the workflow version of python to 3.9, since that's the version we're
using on Heroku.