fix: make YAML reload based on file timestamps (#932)
Loading YAML files can take a long time, which is why we cache them.

We used to have two settings:

- Cache aggressively: best performance, but you have to restart the app
  to see your changes if you change a file (used on Heroku).
- Don't cache: for developers, you can immediately see your changes, but
  reloading a page potentially takes a long time since all YAML files
  are reloaded on every page load.

This PR introduces a bit more clever behavior: automatically reload the
YAML file only if it has changed since the last time it was loaded
(determined by the modification time of the YAML source file). This
makes it do The Right Thing™️ both on Heroku as well as on dev machines,
and so hopefully removes the need for the `NO_DEBUG_MODE` flag
introduced in #817.

(This subjectively makes refreshing a Hedy page locally feel approximately 100x
faster)

This PR also introduces a big refactoring: instead of `load_yaml()`
calls everywhere, introduces the concept of a `YamlFile` object. This
file represents the contents of the indicated file, and can be accessed
as if it was a dictionary with the YAML file contents in it,
automatically reloading as necessary. So instead of:

    @app.route(...)
    def handler():
        my_file = load_yaml('my/yaml/file.yaml')
        return my_file['data']

We do the following:

    my_file = YamlFile.for_file('my/yaml/file.yaml')

    @app.route(...)
    def handler():
        return my_file['data']

The `my_file` object always represents the data that is this YAML file.

Co-authored-by: Felienne <felienne@gmail.com>