Splitting level1.lark file (#1195)
* Splitting level1.lark file

The keywords, which were originally placed in level1.lark, can now be accessed from keywords-en.lark and keywords-nl. This can be used to add multiple languages for the keywords.

* Forgot to remove equal signs, comma's etc  from keyword files

* Resolving comments on pull requests

The 'create_grammar' function will now differentiate between keyword languages. Also, in the 'get_parser' function, the key which is passed into the parser cache depends on the language of the keywords.

* Total grammar saving now accounts on language

The function 'save_total_grammar_file' now takes language in account.

Co-authored-by: Felienne <felienne@gmail.com>