Adds 'start tutorial' button if a tutorial exists for the current lang / level combination (#3242)
**Description**
In this PR we built upon the implementation of a tutorial template. We add a start tutorial button if there is a tutorial found for the current language / level combination. When the tutorial is added to the corresponding YAML it should all work automatically. However, when wanting to re-locate the popup, highlight or hide specific elements new functions should be added and called in the corresponding .ts files in the `/static/js/tutorials` folder.

**Fixes**
Not really a related issue, however we are closer to fixing #2777.

**How to test**
Verify that there is a tutorial button showing if the corresponding tutorials YAML contains data for the current level. This is already the case for the English YAML where we added a dummy entry. **This value should be removed after testing**.

**Screenshot**
![Schermafbeelding 2022-08-30 om 14 09 05](https://user-images.githubusercontent.com/38155424/187433471-f4c15a0e-560d-4fcf-ab82-29219c70adc7.png)