Sort research papers by year in descending order (#3366)
First, provide a one line summary of your changes in the **Title of the PR**

**Now fill out the remainder of this template by replacing all _italic_ content**

**Description**

This PR sorts the research papers by year in descending order. It also puts the year it was published in before the research paper's title.

**Fixes #3365**

Always link the number of the issue or of the discussion that your pr concerns.

**How to test**

* If this is a UI change: 

     * Go to http://localhost:8080/learn-more
     * Scroll down to the `Hedy Research` section

<img width="1148" alt="Screen Shot 2022-10-02 at 6 55 26 PM" src="https://user-images.githubusercontent.com/46979193/193480351-09b9787a-921c-472b-bcda-e72489bd30bc.png">

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [x] Describes changes in the format above (present tense)
- [x] Links to an existing issue or discussion 
- [x] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!