Adds support for comments in levels 2-4 (#3391)
**Description**

Comments are already working in level 1 and in levels 5 and up. This fixes it for the remaining levels and adds some tests. The fix is a grammar update so the # is not part of text, and there are updated tests too.

**Fixes #3286 **

Always link the number of the issue or of the discussion that your pr concerns.

**How to test**
The code below is basically what's used to report the issue and doesn't work correctly in levels 2-4 (Note the single quotes to get correct statements for level 4). You can use this to verify that the issue does not exist on the feature branch.

```
print 'hello world!' # with a comment
print 'hello world!' # with a comment
print 'hello world!' # with a comment
print 'hello world!' # with a comment
print 'hello world!' # with a comment
```

The code base now has tests for comments for all levels.

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [x] Describes changes in the format above (present tense)
- [x] Links to an existing issue or discussion 
- [x] Has a "How to test" section