[UI] Add number of lines to explore and programs page (#3490)
**Description**

Displays the number of lines for stored programs in the explore and programs page.

![image](https://user-images.githubusercontent.com/45865185/197912167-904acc6d-c9e7-4b9f-a1e1-2603f6187a62.png)

![image](https://user-images.githubusercontent.com/45865185/197912683-22ab2368-022b-436d-aa87-1dc9fb467c6f.png)

**Fixes #3339 **

Always link the number of the issue or of the discussion that your pr concerns.

**How to test**

Go to /explore page and /programs page

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Describes changes in the format above (present tense)
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!