[Language] Changing to float division from level 12 on (#3519)
**Description**

Issue #3468 says that it is expected for division to return float numbers starting from level 12. This PR fixes that by changing the operator that translates / from // to / . 

This PR is a draft because I wanted to discuss more deeply about this change, since several tests tell me that this was expected behavior, this tests being:
1. `test_level_12.py::test_int_calc`
2. `test_level_12.py::test_nested_int_calc`

**Fixes #3468 **

**How to test**

I added and changed existing tests.

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Describes changes in the format above (present tense)
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!