Fixes support for pa_Arab_pk (#3599)
**Description**
This is a bit of a work-around to get both Weblate and Babel happy with our translation naming convention. The issue is that Weblate expects the language key to be `pa_PK` whereas Babel expects `pa_Arab_PK`. In this PR I propose the following solution:

- Add a custom locale to Babel -> Redirection `pa_PK` to `pa_Arab_PK`
- Overwrite the internal Babel searching to `pa_Arab_PK` as well
- **But**: Don't do this for the actual `local_parse` file! This should remain `pa_PK`

**Fixes**
This PR fixes #3593.

**How to test**
Verify that Babel works with `pa_PK` while it original didn't. You can verify by editing the .po file in the corresponding translations folder, compile and notice the change in the translated string.