[FIX] Manually adds support for Tswana (#3603)
**Description**
In this PR we fix the support for Tswana that is not natively supported by Babel. We use the same work-around as we used in #3599 for adding support for `pa_PK`. Simply re-direct Babel to another language when the non-supported language is called but **do not** overwrite the locale parser -> keep this to the original "unsupported" language.

**Fixes**
This PR fixes #3600.

**How to test**
Verify that Tswana is listed in the language dropdown selector. Also verify that the translations are wrapped correctly by both changing some stuff in the `tn.yaml` files in `/content` and the .po files in `/translations`. In the second case don't forget to compile the translations.