[FIX] Adds support for adding/removing numbers and list accesses to/from lists (#3293) (#3685)
[FIX] added support for adding/removing numbers and list accesses to/from lists (#3293)


**Description**

Adds support to levels 3+ for adding integers or floats to lists, and removing integers of floats from lists. Also fixes the issue where accessing a list item with square bracket indexing gets changed into a literal string when adding or removing from a list (ex. add list1[2] to list2, appends 'list1[2]' as a string to list2)

**Fixes Issue #3293**

**How to test**

Adds a test to test_level_03.py for checking that a number can be added to a list
Adds tests to test_level_16.py for checking that square bracket list indexing correctly gets added and removed from a list