[REFACTORING] Change Parsons YAML structure for code lines (#3697)
**Description**
In this PR we change the structure of the Parsons code_lines in the YAML to make translating easier. Instead of a dict structure we change this to a whole snippet that is splitted on the front-end on runtime. We keep support for both `code_lines` and `code` as a yaml attribute for now to support relic YAML files. This should be cleaned-up once all translations and weblate are happy.

**Fixes**
This PR fixes #3677.

**How to test**
All should work as before as this is only a refactoring.