[FIX] Refactor 'explore' page (#3875)
Refactor the database access of the /explore page. We now use key-only indices to filter the database records (potentially still not using the most effective database access mechanism, but should be better), and use batch get to retrieve the public profile information. Also put a timeout on it.

All this effort for nothing -- turns out the time is mostly spent in translating the keywords in the examples. This invokes at least one and potentially 2 Hedy parses for every program on the explore page.

Still committing this because it should be better, but the real work is elsewhere.

## Attempt 2

After talking about this with Felienne, we are now getting rid of the language choice on the explore page. The upshot is that we don't need do to translations anymore, and so we save 1 or 2 parses!

Also in this PR:

- Print querylog records in debug mode, so they're easier to see and debug
- Fix an issue in the querylog records where querystrings are incorrectly formatted
- Add a `safe_format` function which will not throw `KeyError` when provided values don't match the placeholders.
- Add `try_parse_int()` which will not throw `ValueError` when the given input does not parse cleanly.
- Make the `/explore` page automatically submit a new search when the dropdown values are changed (users don't need to explicitly click "Search" anymore).
- Make the default dropdown values in the `/explore` page selectable and make them count as "no filter" (users don't need to click the "Reset" button anymore).
- Turned all the `0: 0`, `1: 1` etc entries in the keyword files into `d0: 0`, `d1: 1` etc, because the names of the entries that go into `string.format()` (and `safe_format`) may not be numbers.
- Fix the `start-debug-server` script to actually properly record the PID of the Python process


**How to test**

The explore page should look and behave the same.