[CHORE] Design: move teacher nav to top bar (#3912)
Move the teacher navigation out of the profile dropdown, into the top bar, where it will be more prominently visible.

Also make the teacher's manual publicly visible, not only to teachers, so that teachers (finding the link via the colophon) can have a preview at the learning material that's available to them, hopefully enticing them to register.

## After

![image](https://user-images.githubusercontent.com/524162/210543663-1fde9ff7-c7c3-4b8a-9133-f056d7ca0e95.png)