[FIX] Improve logout experience (#4157)
Previously, when you log out, you would be shown a JSON error message (because the browser will send `Accept: */*`, which includes JSON, and that's what we feature tested for). The goal was to respond with JSON to the Ajax requests. Fix by feature testing for yes JSON, but not HTML.

Also, the previous PR made it so that if you log out, you stay on the same page. But if this page is now a "my profile" page or "programs" page, it shows you a very big unfriendly "you are not logged in message".

Add a new `@requires_login_redirect` decorator, which redirects you back to the front page if you end up not logged in on certain pages.

**How to test**

Log in, go to your profile page, then log out. You end up back at the front page.