[TESTING] Weblate fix script (#4544)
One of our biggest time sinks is fixing Weblate snippets (it is mine for sure and I could use my time in better ways). Sometimes, this is needed because a weblate PR is blocking other content PRs. While of course it is not great to throw away snippets, and we should fix them when possible, this PR introduces an environment variable called `fix_for_weblate`, which can be set to revert broken snippets back to English. So if needed this can be used