[REFACTOR] Faster CI (#4574)
## Description

Fixes #4546 in a fundamentally improved way in 3 parts, plus one part with some suggestions for the future.

This speeds up the unit tests:
- non-cached run: was 1.5 hours -> now roughly 35 minutes https://github.com/jtwaleson/hedy/actions/runs/6371719946/attempts/1
- cached run: was 11 minutes -> now roughly 7 minutes https://github.com/jtwaleson/hedy/actions/runs/6371719946/attempts/2

Follow-up work to split the work out into multiple jobs is underway, reducing the total time to 7 minutes uncached and 3 minutes cached. I'll create a separate PR for that once this is merged, as it is dependent on this work, but I don't want to make the PR too big.

**Part 1) using a github cache**
   
We previously used a snippet hash pickled file to see if snippets had
been executed successfully already on this version of hedy.

This was fine but had two downsides:
- the snippet file changed frequently and was checked into git, leading
  to frequent non-functional changes
- using multiple processes to speed up the CI run was not possible
  as each process would try to overwrite the snippet hash file

Now we introduce pytest-xdist to speed up the test runs by using 2
processes concurrently. We can attempt to go to 4 but the Github runners
have 2 cores and can go out of memory if we go to more concurrency.

We also use a directory based cache instead of pickling a set. We use
concurrency-friendly OS primitives to see if a file exists (open and 
os.path.isfile). This way multiple processes can read/write the cache
at the same time.

This cache is .gitignore'd so it's not stored in the repository. It is
however stored as a github cache entry. The key uses the hedy language
definition as an input and then hashes it with sha256sum. This means that
any change to the hedy files will invalidate the cache, which is what we
want.

This removes the need for the periodically generating of the snippet
hashes, so that workflow is removed.

As the entire hedy language definition was calculated with every test
run, and this was hashed with md5. We change this a bit to use the more
modern and standard but still fast sha1. The sha1 for the entire
language is only used once to calculate a short digest for the language.
That digest is then concatenated to the snippets to form the individual
test snippets. This shaves off only a couple of milliseconds and a
couple of KB of memory usage, but it's nice.

**Part 2) using translation key caching**

With every single unit test that translated error messages this function
was hit, which would parse the yaml files from disk. A simple cache
makes this 10x as fast, shaving about 15 seconds from the CI run.

When we work on the language locally and introduce new keywords, we
might now need to restart the process, but I believe this should
already be covered by flask.

**Part 3) use functools.cache for parser cache**

Will pick this up in a future PR.

~~We used custom cache for the parsers. This was fine but we can use
out-of-the-box tooling just as well.~~

~~functools.cache was introduced in python 3.9 and is a nice one-liner.~~

~~However, after this, unit tests were failing as the cache key was
dependent on a "hidden" variable, sourcemap.skip_faulty.~~
    
~~This also makes that relationship transparent.~~

**Part 4) Added some TODOs for future improvements**

We run some tests sequentially which could be done with a `matrix` approach in github actions. This way they run parallel.

## How to test

- Thorough code review
- Observe CI test runs and see the improvement

## Checklist

- [x] Contains one of the PR categories in the name
- [x] Describes changes in the format above
- [x] Links to an existing issue or discussion
- [x] Has a "How to test" section