[LANGUAGE, FIX] Fix non-en colors in lists and ask (#4756)
Reverts hedyorg/hedy#4754. This is a redo of #4675, which fixed https://github.com/hedyorg/hedy/issues/3356 but caused #4752.

This has additional commits that also fix #4752.

**Original text from #4675:**

**Description**

Allows the use of Dutch (and other langs) names in a list, fixing  #3356. This new way of processing numbers to English at runtime also allows for using non-English colors in an ask:

<img width="1608" alt="image" src="https://github.com/hedyorg/hedy/assets/1003685/fa373815-47d5-4e68-b50b-d69404082c6f">

This PR:
<img width="1592" alt="image" src="https://github.com/hedyorg/hedy/assets/1003685/07d0f3ee-e8aa-4407-ad5d-0660c651cba8">

main:
<img width="1422" alt="image" src="https://github.com/hedyorg/hedy/assets/1003685/c7a2d2e5-7c33-48a7-9b63-ce9b0182d932">

**Fixes #3356**

**How to test**

I added a test about the list access, for the ask issue, check that the images above work for you also.