🪲 Allow list access to appear on lhs of and/or conditions (#5410)
Fixes #5398 
In level 17, the OR and AND conditions could not have a list access on the left hand side. The problem occurs because the `]` has optional space after it while the `and/or` operator has mandatory space before it. When parsing `if 'red' == colors[1] and ...`, the parser matches the space between `]` and `and` as the optional space after `]` and not as the mandatory space before `and`.

The solution in this PR relaxes the constraint in the `and` and `or` operators by making the space before them optional.

**How to test**
- Confirm that the following program works correctly in level 17:
```
f = ['.', '.', '.']
if f[1] == f[2] and f[2] == f[3]
    print 'test'
```