💻 add roles of variables: stepper in for loops (#5538)
I've added the stepper role, it works with for loops, but other kinds of steppers are not working yet. 

Also started using labels so that the variable role names can be translated into the different languages (added the English myself, with container now called list and input instead of input constant).

**How to test**

Follow these steps to verify this PR works as intended:

* change the feature flag showRoles in debugging.ts to true
* go to Hedy level 11 and run this program (or something similar) in the debugger:
```
for counter in range 1 to 10
    print counter
print 'Ready or not. Here I come!'
```
* check that in the variables view, the role that is shown for the variable counter is stepper. 

**Checklist**
Done? Check if you have it all in place using this list: (mark with x if done)

- [ x] Contains one of the PR categories in the name
- [ x] Describes changes in the format above
- [ ] Links to an existing issue or discussion
- [ x] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!