🪲 Extend ask_error_nq to include ask with = (#5736)
Fixes #3727
If the `ask` command is used with a missing end quote in levels 4 and above, we return an unquoted error. However, if the user used `=` instead of `is`, the error is not displayed and the code is interpreted as an assignment.

With this PR, the behavior of the `ask` command with an `=` and `is` is made symmetrical - the unquoted error is shown in both cases. Note that this revealed a couple of errors in the translated content.

**How to test**
In level 6 (when the `=` is also available) try the following program:
```
a = ask 'there is a missing closing quote
print a
```
You should get an error about missing quotation marks. You should not get the output `ask 'there is a missing closing quote`.